Separate out GPA map, params and directives
Perform a general tidy up of the code derived from C, making it a
more rust-native. The memory map is now buillt as a separate
structure making it easy to organise the order of the directives.
This allows the directives to be assembled in the same order as
the current C builder meaning it can be used to verify the output
of this builder.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>